Add entry points execution feature, and shrink menu
Result of entry points executions are displayed in a formatted JSON5 view.
This is a simplified JSON format which omits quotes whenever possible.

Menu was too big and took too much space. We reduced it and used tabstrip
for the sections. We also added an empty section that will be implemented
later: "Loaders".