Create a manager to retrieve Brayns host:port
This info is taken from the URL argument `host`.
But if it is missing, the user is prompted to enter the host:port manually.

This manager is instantiated only by the global factory.
This is a common pattern of the SOLID approach.
This way, the rest of the application does not have to worry about the dependencies
of the modules implementations. Replacing a module with a mock for testing is facilitated.